+++ src/MoesTavern.Api/Schema/InventoryQuery.cs
@@ -6,29 +6,31 @@ using System;
 
 namespace MoesTavern.Api.Schema
 {
-    public interface IInventoryQuery : IObjectGraphType 
-    {}
-    
+    public interface IInventoryQuery : IObjectGraphType
+    { }
+
     public class InventoryQuery : ObjectGraphType, IInventoryQuery
     {
         public InventoryQuery(IHttpContextAccessor contextAccessor, IInventoryRepository inventory)
         {
             FieldAsync<ListGraphType<BeerGraphType>>(
                name: "whatsOnTap",
-               description: "Retuns all beers in your inventory",
-               resolve: async context => {
+               description: "Returns all beers in your inventory",
+               resolve: async context =>
+               {
                    //TODO: implement
                    throw new NotImplementedException();
-                });
+               });
 
             FieldAsync<BeerGraphType>(
                name: "findBeer",
                description: "Finds the beer you're looking for",
                arguments: new QueryArguments(new QueryArgument<IntGraphType> { Name = "id" }),
-               resolve: async context => {
+               resolve: async context =>
+               {
                    //TODO: implement
                    throw new NotImplementedException();
-                });
+               });
         }
     }
 }
